Implement push to server

+++ onion-bt-wardriving.go
@@ -3,9 +3,11 @@ package main
 import (
 	"bytes"
 	"encoding/json"
+	"flag"
 	"fmt"
 	"github.com/peterbourgon/diskv"
 	"github.com/xperimental/onion-weather/oled"
+	"net/http"
 	"os"
 	"os/exec"
 	"os/signal"
@@ -21,30 +23,43 @@ type device struct {
 	LastSeen int64
 }
 
+type deviceFlat struct {
+	device
+	Mac string
+}
+
 var re = regexp.MustCompile("(?im)^[^0-9a-f]*((?:[0-9a-f]{2}:){5}[0-9a-f]{2})\\s*([^\\s].*)?$")
-var displayBuffer = make([]string, 8, 8)
+var displayBuffer = make([]string, 8)
 var display oled.Display
 
 var dv *diskv.Diskv
 
 func main() {
+	dataEndpoint := flag.String("push-to-server", "", "endpoint to send collected data to")
+	flag.Parse()
+
 	setupPersistence()
-	setupBt()
-	setupOled()
-	defer display.Close()
 
-	sig := make(chan os.Signal, 1)
-	signal.Notify(sig, os.Interrupt, os.Kill, syscall.SIGTERM)
-	defer signal.Stop(sig)
+	if *dataEndpoint != "" {
+		sendAllToEndpoint(*dataEndpoint)
+	} else {
+		setupBt()
+		setupOled()
+		defer display.Close()
 
-	for {
-		select {
-		case <-time.After(1 * time.Millisecond):
-			loop()
-		case s := <-sig:
-			fmt.Println("Got signal:", s)
-			fmt.Println("Quitting...")
-			return
+		sig := make(chan os.Signal, 1)
+		signal.Notify(sig, os.Interrupt, os.Kill, syscall.SIGTERM)
+		defer signal.Stop(sig)
+
+		for {
+			select {
+			case <-time.After(1 * time.Millisecond):
+				loop()
+			case s := <-sig:
+				fmt.Println("Got signal:", s)
+				fmt.Println("Quitting...")
+				return
+			}
 		}
 	}
 }
@@ -230,4 +245,62 @@ func notify() {
 	if err != nil {
 		fmt.Println(err)
 	}
+}
+
+func sendAllToEndpoint(endpoint string) {
+	devices := make(chan deviceFlat)
+	go collectEntries(devices)
+
+	sendDone := make(chan error)
+
+	counter := 0
+	buffer := make([]deviceFlat, 20, 20)
+	for d := range devices {
+		// Collect 20 devices and send them out -> and repeat
+		if counter == 20 {
+			go sendToEndpoint(endpoint, buffer, sendDone)
+			counter = 0
+		}
+		buffer[counter] = d
+		counter++
+	}
+
+}
+
+func sendToEndpoint(endpoint string, devices []deviceFlat, done chan error) {
+	defer close(done)
+	data, err := json.Marshal(devices)
+	if err != nil {
+		fmt.Printf("Error during json marshal %v", err)
+		done <- err
+		return
+	}
+	_, err = http.Post(endpoint+"/data", "application/json", bytes.NewBuffer(data))
+	if err != nil {
+		fmt.Printf("Error posting data to %s: %v", endpoint+"/data", err)
+		done <- err
+		return
+	}
+}
+
+func sendDoneSignal(endpoint string) {
+	_, err := http.Get(endpoint + "/done")
+	if err != nil {
+		fmt.Printf("Error signaling done to %s: %v", endpoint+"/done", err)
+	}
+}
+
+func collectEntries(devices chan deviceFlat) {
+	cancel := make(chan struct{})
+	c := dv.Keys(cancel)
+	for mac := range c {
+		deviceData := readDevice(mac)
+		if deviceData == nil {
+			fmt.Printf("Failed to read device with key %s\n", mac)
+		}
+		device := deviceFlat{Mac: mac}
+		device.device = *deviceData
+		devices <- device
+	}
+	close(devices)
 }
